Move Like/Dislike/Attend add/remove code to include/like.php
